+++ tools/src/librw11/Rw11CntlRK11.cpp
@@ -1,4 +1,4 @@
-// $Id: Rw11CntlRK11.cpp 1089 2018-12-19 10:45:41Z mueller $
+// $Id: Rw11CntlRK11.cpp 1090 2018-12-21 12:17:35Z mueller $
 //
 // Copyright 2013-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 // Other credits: 
@@ -15,6 +15,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-19  1090   2.0.9  use RosPrintf(bool)
 // 2018-12-15  1082   2.0.8  use std::bind or lambda instead of boost
 // 2018-12-09  1080   2.0.7  use HasVirt(); Virt() returns ref
 // 2018-10-28  1062   2.0.6  replace boost/foreach
@@ -310,8 +311,8 @@ void Rw11CntlRK11::Dump(std::ostream& os, int ind, const char* text,
   os << bl << "  fRd_addr:        " << fRd_addr << endl;
   os << bl << "  fRd_lba:         " << fRd_lba  << endl;
   os << bl << "  fRd_nwrd:        " << fRd_nwrd << endl;
-  os << bl << "  fRd_fu:          " << fRd_fu  << endl;
-  os << bl << "  fRd_ovr:         " << fRd_ovr  << endl;
+  os << bl << "  fRd_fu:          " << fRd_fu   << endl;
+  os << bl << "  fRd_ovr:         " << RosPrintf(fRd_ovr)  << endl;
   fRdma.Dump(os, ind+2, "fRdma: ", detail);
   Rw11CntlBase<Rw11UnitRK11,8>::Dump(os, ind, " ^", detail);
   return;