+++ tools/src/librw11/Rw11UnitTape.cpp
@@ -1,4 +1,4 @@
-// $Id: Rw11UnitTape.cpp 1080 2018-12-09 20:30:33Z mueller $
+// $Id: Rw11UnitTape.cpp 1090 2018-12-21 12:17:35Z mueller $
 //
 // Copyright 2015-2018 by Walter F.J. Mueller <W.F.J.Mueller@gsi.de>
 //
@@ -13,6 +13,7 @@
 // 
 // Revision History: 
 // Date         Rev Version  Comment
+// 2018-12-19  1090   1.0.3  use RosPrintf(bool)
 // 2018-12-09  1080   1.0.2  use HasVirt(); Virt() returns ref
 // 2017-04-07   868   1.0.1  Dump(): add detail arg
 // 2015-06-04   686   1.0    Initial version
@@ -25,6 +26,9 @@
 */
 
 #include "librtools/Rexception.hpp"
+
+#include "librtools/RosFill.hpp"
+#include "librtools/RosPrintf.hpp"
 
 #include "Rw11UnitTape.hpp"
 
@@ -238,8 +242,8 @@ void Rw11UnitTape::Dump(std::ostream& os, int ind, const char* text,
   RosFill bl(ind);
   os << bl << (text?text:"--") << "Rw11UnitTape @ " << this << endl;
   os << bl << "  fType:           " << fType  << endl;
-  os << bl << "  fEnabled:        " << fEnabled << endl;
-  os << bl << "  fWProt:          " << fWProt << endl;
+  os << bl << "  fEnabled:        " << RosPrintf(fEnabled) << endl;
+  os << bl << "  fWProt:          " << RosPrintf(fWProt) << endl;
   os << bl << "  fCapacity:       " << fCapacity << endl;
 
   Rw11UnitVirt<Rw11VirtTape>::Dump(os, ind, " ^", detail);